fastapi_serving/requirements.txt
fastapi
uvicorn
numpy
onnxruntime-gpu
